pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "url_shortener"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.27.1"
fastapi = "^0.109.2"
flake8 = "^7.0.0"
jinja2 = "^3.1.3"
validators = "^0.22.0"
asyncio = "^3.4.3"
requests = "^2.31.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
pyshorteners = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
